Renamed kernelLaunch to kabob case
